// server/src/data/types.d.ts
export type DBCard = {
  Z_PK: number;
  Z_ENT: number;
  Z_OPT: number;
  ZTCGPLAYERID: number;
  ZCONVERTEDMANACOST: number;
  ZCREATIONDATE: 697776984.703475;
  ZPRICE: number;
  ZCOLLECTORNUMBER: string;
  ZFACEARTIST: string;
  ZFACEMANACOST: string;
  ZFACENAME: string;
  ZFACETYPELINE: string;
  ZORACLEID: string;
  ZPARSEDCOLORCOST: string;
  ZPARSEDCOLORIDENTITY: string;
  ZPARSEDEXTRAS: string;
  ZPARSEDLANGUAGE: string;
  ZPARSEDRARITY: string;
  ZPARSEDSETCODE: string;
  ZPARSEDSETNAME: string;
  ZPRICESKU: string;
  ZSCRYFALLID: string;
  ZBINDERID: number | null;
  ZID: Buffer;
  ZLISTID: number | null;
};